+++ packages/core/src/index.ts
@@ -7,10 +7,3 @@ export type {
 export { executeRequest } from './http-client.js';
 export { getAiCoreDestination } from './context.js';
 export { OpenApiRequestBuilder } from './openapi-request-builder.js';
-export type {
-  AzureOpenAiChatModel,
-  AzureOpenAiEmbeddingModel,
-  GcpVertexAiChatModel,
-  AwsBedrockChatModel,
-  ChatModel
-} from './model-types.js';
