+++ packages/core/src/internal.ts
@@ -1,4 +1,3 @@
 export * from './context.js';
 export * from './http-client.js';
 export * from './openapi-request-builder.js';
-export * from './model-types.js';
