+++ packages/foundation-models/src/openai/openai-embedding-client.ts
@@ -1,12 +1,9 @@
-import {
-  type CustomRequestConfig,
-  type AzureOpenAiEmbeddingModel,
-  executeRequest
-} from '@sap-ai-sdk/core';
+import { type CustomRequestConfig, executeRequest } from '@sap-ai-sdk/core';
 import {
   getDeploymentId,
   type ModelDeployment
 } from '@sap-ai-sdk/ai-api/internal.js';
+import type { AzureOpenAiEmbeddingModel } from '../model-types.js';
 import type {
   OpenAiEmbeddingParameters,
   OpenAiEmbeddingOutput
