+++ packages/orchestration/src/index.ts
@@ -39,3 +39,11 @@ export { OrchestrationClient } from './orchestration-client.js';
 export { azureContentFilter } from './orchestration-filter-utility.js';
 
 export { OrchestrationResponse } from './orchestration-response.js';
+
+export type {
+  AzureOpenAiChatModel,
+  AicoreOpensourceChatModel,
+  GcpVertexAiChatModel,
+  AwsBedrockChatModel,
+  ChatModel
+} from './model-types.js';
