--- tests/type-tests/test/model.test-d.ts
@@ -1,5 +1,0 @@
-import { ChatModel, AzureOpenAiChatModel } from '@sap-ai-sdk/core';
-
-expect<ChatModel>('custom-model');
-expect<AzureOpenAiChatModel>('custom-model');
-expect<ChatModel>('gpt-4-32k');
